mod-ur: rename midi-test.ts to midi-core.ts to avoid test conflict
